adds new entity Goal to CoreData (xcdatamodeld)
while trying to instantiate it inside of GoalsVC, we first had
to assign a `Core Data Model Identifier` on the `Identity and Type`
panel of the Goal entity entry. Without that, it was not possible
to have it appear in the autocompletion from XCode, even after
restarting it.

Maybe it is related to the versions of Swift (6.0) and Xcode I'm
using (16.2) which, compared to the ones used in the Coursera
course (Swift 3 and Xcode 9/10), it's way newer